Removed IBasicComponentCollection and renamed the taggable variant to IComponentCollection.
